+++ internal/engine/index/impl/gpu/gamma_index_ivfpq_gpu.h
@@ -83,7 +83,7 @@ class GammaIVFPQGPUIndex : public IndexModel {
   long GetTotalMemBytes() override;
 
   Status Dump(const std::string &dir) override;
-  Status Load(const std::string &index_dir, int &load_num) override;
+  Status Load(const std::string &index_dir, int64_t &load_num) override;
 
  private:
   int GPUThread();
