+++ internal/engine/index/index_io.h
@@ -146,5 +146,5 @@ int WriteInvertedLists(faiss::IOWriter *f,
                        realtime::RTInvertIndex *rt_invert_index);
 Status ReadInvertedLists(faiss::IOReader *f,
                          realtime::RTInvertIndex *rt_invert_index,
-                         int &indexed_vec_count);
+                         int64_t &indexed_vec_count);
 }  // namespace vearch
